+++ XCDYouTubeKit Tests/XCDYouTubeClientTestCase.m
@@ -3,8 +3,11 @@
 //
 
 #import "XCDYouTubeKitTestCase.h"
-
+#if TARGET_OS_OSX
+#import <AppKit/NSImage.h>
+#endif
 #import <XCDYouTubeKit/XCDYouTubeClient.h>
+#import <XCDYouTubeKit/XCDYouTubeVideoOperation.h>
 
 @interface XCDYouTubeClientTestCase : XCDYouTubeKitTestCase
 @end
@@ -31,9 +34,75 @@
 		XCTAssertNil(error);
 		XCTAssertEqualObjects(video.identifier, @"9TTioMbNT9I");
 		XCTAssertEqualObjects(video.title, @"Super Mario Bros Theme Song on Wine Glasses and a Frying Pan (슈퍼 마리오 브라더스 - スーパーマリオブラザーズ - 超級瑪莉)");
-		XCTAssertNotNil(video.thumbnailURL);
+		XCTAssertEqualObjects(video.author, @"Dan Newbie");
+		XCTAssertEqualObjects(video.channelIdentifier, @"UCxjo61fSS-hR7RiyBxnnHdg");
+		XCTAssertNotEqualObjects(video.videoDescription, @"");
+		XCTAssertNotNil(video.thumbnailURLs.firstObject);
 		XCTAssertTrue(video.streamURLs.count > 0);
 		XCTAssertTrue(video.duration > 0);
+		[expectation fulfill];
+	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+#if TARGET_OS_OSX
+- (void) testThatThumbnailsAreOrderedCorrectly
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"9TTioMbNT9I" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNil(error);
+		XCTAssertNotNil(video.thumbnailURLs);
+		XCTAssertNotEqual(video.thumbnailURLs.count, 0);
+
+		NSMutableArray<NSImage *>*thumbnailImages = [NSMutableArray new];
+
+		for (NSURL *thumbnailURL in video.thumbnailURLs)
+		{
+			[thumbnailImages addObject:[[NSImage alloc]initWithContentsOfURL:thumbnailURL]];
+		}
+
+		[thumbnailImages sortUsingComparator:^NSComparisonResult(NSImage * _Nonnull image1, NSImage *  _Nonnull image2)
+		{
+			BOOL isSmaller = image1.size.width < image2.size.width && image1.size.height < image2.size.height;
+			XCTAssertTrue(isSmaller, @"`thumbnailURLs` should be ordered from smallest to largest");
+			return isSmaller;
+		}];
+
+		[expectation fulfill];
+	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+#endif
+- (void) testThatVideoHasOtherStreams
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"sBHFOh5qe20" completionHandler:^(XCDYouTubeVideo *mainVideo, NSError *error)
+	{
+		XCTAssertNil(error);
+		XCTAssertNotNil(mainVideo.videoIdentifiers);
+		
+		NSMutableArray *operations = [NSMutableArray new];
+		NSOperationQueue *queue = [NSOperationQueue new];
+		queue.maxConcurrentOperationCount = 6;
+		
+		for (NSString *videoIdentifier in mainVideo.videoIdentifiers)
+		{
+			[operations addObject:[[XCDYouTubeVideoOperation alloc]initWithVideoIdentifier:videoIdentifier languageIdentifier:nil]];
+		}
+		
+		XCTAssertTrue(operations.count != 0);
+		[queue addOperations:operations waitUntilFinished:YES];
+		
+		for (XCDYouTubeVideoOperation *operation in operations)
+		{
+			XCTAssertNil(operation.error);
+			XCTAssertNotNil(operation.video);
+			XCTAssertTrue(operation.video.streamURLs.count > 0);
+			XCTAssertTrue(operation.video.duration > 0);
+			XCTAssertNotEqualObjects(operation.video, mainVideo, @"None of the `videoIdentifiers` returned from the `mainVideo` should be the same `videoIdentifier` was the `mainVideo`");
+		}
+		
 		[expectation fulfill];
 	}];
 	[self waitForExpectationsWithTimeout:5 handler:nil];
@@ -103,8 +172,9 @@
 	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video.title);
+		XCTAssertTrue(video.viewCount > 0);
 		XCTAssertNotNil(video.expirationDate);
-		XCTAssertNotNil(video.thumbnailURL);
+		XCTAssertNotNil(video.thumbnailURLs.firstObject);
 		XCTAssertTrue(video.streamURLs.count > 0);
 		XCTAssertTrue(video.duration > 0);
 		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop) {
@@ -122,7 +192,8 @@
 	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video.title);
-		XCTAssertNotNil(video.thumbnailURL);
+		XCTAssertTrue(video.viewCount > 0);
+		XCTAssertNotNil(video.thumbnailURLs.firstObject);
 		XCTAssertNotNil(video.streamURLs[XCDYouTubeVideoQualityHTTPLiveStreaming]);
 		[expectation fulfill];
 	}];
@@ -138,7 +209,8 @@
 	 {
 		 XCTAssertNil(error);
 		 XCTAssertNotNil(video.title);
-		 XCTAssertNotNil(video.thumbnailURL);
+		 XCTAssertTrue(video.viewCount > 0);
+		 XCTAssertNotNil(video.thumbnailURLs.firstObject);
 		 XCTAssertTrue(video.streamURLs.count > 0);
 		 XCTAssertTrue(video.duration > 0);
 		 [video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop) {
@@ -158,8 +230,9 @@
 	{
 		XCTAssertNil(error);
 		XCTAssertNotNil(video.title);
+		XCTAssertTrue(video.viewCount > 0);
 		XCTAssertNotNil(video.expirationDate);
-		XCTAssertNotNil(video.thumbnailURL);
+		XCTAssertNotNil(video.thumbnailURLs.firstObject);
 		XCTAssertTrue(video.streamURLs.count > 0);
 		XCTAssertTrue(video.duration > 0);
 		NSMutableURLRequest *request = [NSMutableURLRequest requestWithURL:video.streamURLs[@(XCDYouTubeVideoQualityMedium360)]];
@@ -171,6 +244,371 @@
 		}];
 		[dataTask resume];
 	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testVideo1ReturnsSomePlayableStreams
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	
+	//These are the playble itag stream for `cdqP6wI8TCc` as of Feb 12, 2020 in the US
+	NSArray<NSNumber *>*playableStreamKeys = @[@140, @136, @251, @134];
+	
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"cdqP6wI8TCc" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+		
+		[[XCDYouTubeClient defaultClient]queryVideo:video cookies:nil completionHandler:^(NSDictionary * _Nonnull streamURLs, NSError * _Nullable queryError, NSDictionary<id, NSError *> *streamErrors) {
+			
+			XCTAssertNil(queryError);
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+			
+			for (NSNumber *itag in playableStreamKeys)
+			{
+				XCTAssertTrue([streamURLs.allKeys containsObject:itag]);
+			}
+			
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+			
+			XCTAssertEqual(playableStreamKeys.count, streamURLs.count, @"`streamURLs` count should be equal to `playableStreamKeys` count");
+			XCTAssertNotEqual(video.streamURLs.count, streamURLs.count, @"`streamURLs` count should not be equal since this video contains some streams are unplayable");
+			
+			[expectation fulfill];
+		}];
+	}];
+	
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+// Disable internet connection before running to allow some queries to fail
+// Also, this test requires using Charles Proxy tools (or similar app) to block some of the streamURLs
+- (void) testVideo1ReturnsSomePlayableStreamsEvenIfSomeFailDueToConnectionError_offline
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"cdqP6wI8TCc" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+		
+		[[XCDYouTubeClient defaultClient]queryVideo:video cookies:nil completionHandler:^(NSDictionary * _Nonnull streamURLs, NSError * _Nullable queryError, NSDictionary<id, NSError *> *streamErrors) {
+			
+			XCTAssertNil(queryError);
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+			
+			XCTAssertTrue(streamErrors.count != 0);
+			for (NSError *streamError in streamErrors.allValues)
+			{
+				XCTAssertNotNil(streamError.localizedDescription);
+			}
+			
+			XCTAssertNotEqual(video.streamURLs.count, streamURLs.count, @"`streamURLs` count should not be equal since this video contains some streams are unplayable");
+			
+			[expectation fulfill];
+		}];
+	}];
+	
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+// Disable internet connection before running to allow all queries to fail
+- (void) testVideo1ReturnsNoPlayableStreamsBecauseConnectionError_offline
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"cdqP6wI8TCc" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+		
+		[[XCDYouTubeClient defaultClient]queryVideo:video cookies:nil completionHandler:^(NSDictionary * _Nonnull streamURLs, NSError * _Nullable queryError, NSDictionary<id, NSError *> *streamErrors) {
+			
+			XCTAssertNotNil(queryError);
+			XCTAssertNil(streamURLs);
+			XCTAssertTrue(streamErrors.count != 0);
+			XCTAssertTrue([NSThread isMainThread]);
+			
+			for (NSError *streamError in streamErrors.allValues)
+			{
+				XCTAssertNotNil(streamError.localizedDescription);
+			}
+			
+			XCTAssertNotEqual(video.streamURLs.count, streamURLs.count, @"`streamURLs` count should not be equal since this video contains some streams are unplayable");
+			
+			[expectation fulfill];
+		}];
+	}];
+	
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testVideo2ReturnsAllPlayableStreams
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"QcIy9NiNbmo" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+		
+		[[XCDYouTubeClient defaultClient]queryVideo:video cookies:nil completionHandler:^(NSDictionary * _Nonnull streamURLs, NSError * _Nullable queryError, NSDictionary<id, NSError *> *streamErrors) {
+			
+			XCTAssertNil(queryError);
+			XCTAssertNil(streamErrors);
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+			
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+			
+			XCTAssertEqual(video.streamURLs.count, streamURLs.count, @"`streamURLs` count should be equal since all the streams are playable.");
+			
+			[expectation fulfill];
+		}];
+	}];
+	
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testVideo3ReturnsSomePlayableStreams
+{
+	/**
+	 * This video `550S-6XVRsw` contains some streams (e.g. itag=22)  that don't play (the file appeas to be incomplete on YouTube's servers).
+	 * This test ensures that we catch those kinds of errors and they aren't included in the `streamURLs`
+	 * See https://github.com/0xced/XCDYouTubeKit/issues/456 for more information.
+	 */
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	NSNumber *nonPlayableStreamKey = @(XCDYouTubeVideoQualityHD720);
+	
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"550S-6XVRsw" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+		
+		[[XCDYouTubeClient defaultClient]queryVideo:video cookies:nil completionHandler:^(NSDictionary * _Nonnull streamURLs, NSError * _Nullable queryError, NSDictionary<id, NSError *> *streamErrors) {
+			
+			XCTAssertNil(queryError);
+			XCTAssertNotNil(streamErrors);
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+			
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+			
+			XCTAssertNotEqual(video.streamURLs.count, streamURLs.count, @"`streamURLs` count should not be equal since this video contains some streams are unplayable");
+			XCTAssertNil(streamURLs[nonPlayableStreamKey], @"itag 22 should not be available in this stream.");
+			//I noticed when the file stored on the server is not complete we get this error
+			XCTAssertTrue([streamErrors.allValues.firstObject.domain isEqual:NSURLErrorDomain]);
+			XCTAssertEqual(streamErrors.allValues.firstObject.code, NSURLErrorNetworkConnectionLost);
+			XCTAssertNotNil(streamErrors.allValues.firstObject.userInfo[NSLocalizedRecoverySuggestionErrorKey]);
+			XCTAssertTrue([streamErrors.allValues.firstObject.userInfo[NSLocalizedRecoverySuggestionErrorKey] isEqual:@"The file stored on the server may be incomplete."]);
+			[expectation fulfill];
+		}];
+	}];
+	
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testThatQueryingLiveVideoReturnsPlayableStreams
+{
+	/**
+	 * This video `hHW1oY26kxQ` is a live stream
+	 * See https://github.com/0xced/XCDYouTubeKit/issues/456 for more information.
+	 */
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"hHW1oY26kxQ" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+		
+		[[XCDYouTubeClient defaultClient]queryVideo:video cookies:nil completionHandler:^(NSDictionary * _Nonnull streamURLs, NSError * _Nullable queryError, NSDictionary<id, NSError *> *streamErrors) {
+			
+			XCTAssertNil(queryError);
+			XCTAssertNotNil(streamURLs[XCDYouTubeVideoQualityHTTPLiveStreaming], @"Should contain live stream");
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+			
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+
+			[expectation fulfill];
+		}];
+	}];
+	
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testQueryingWithSpecifiedStreamURLs
+{
+	/**
+	 * This video `NZzQQ1090wc` {itag 137 & 22) are reachable
+	 * This test ensures that when specifying streamURLs that are in the `video` object that the operation returns only streamURLs that we specified when complete
+	 */
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"NZzQQ1090wc" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+
+		NSArray<NSNumber *>*specifiedStreamiTags = @[@137, @22];
+		NSMutableDictionary *specifiedStreamURLs = [NSMutableDictionary new];
+
+		for (NSNumber *itag in specifiedStreamiTags)
+		{
+			if (video.streamURLs[itag])
+			{
+				specifiedStreamURLs[itag] = video.streamURLs[itag];
+			}
+		}
+
+		[[XCDYouTubeClient defaultClient] queryVideo:video streamURLsToQuery:specifiedStreamURLs options:nil cookies:nil completionHandler:^(NSDictionary * _Nullable streamURLs, NSError * _Nullable queryError, NSDictionary<id,NSError *> * _Nullable streamErrors)
+		{
+			XCTAssertNil(queryError);
+			XCTAssertNil(streamErrors);
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+
+			XCTAssertEqual(specifiedStreamURLs.count, streamURLs.count, @"`streamURLs` count should be equal since we specified two streams that we know are reachable.");
+
+			[expectation fulfill];
+		}];
+	}];
+
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testQueryingWithSpecifiedStreamURLsSomeNotBeingInVideoObject
+{
+	/**
+	 * This video `NZzQQ1090wc` {itag 137 & 22) are reachable
+	 * This test ensures that when specifying streamURLs (with some not being in `video.streamURLs` that operation completes)
+	 */
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"NZzQQ1090wc" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+
+		NSArray<NSNumber *>*specifiedStreamiTags = @[@137, @22, @1111111];
+		NSMutableDictionary *specifiedStreamURLs = [NSMutableDictionary new];
+
+		for (NSNumber *itag in specifiedStreamiTags)
+		{
+			if ([itag isEqual:@1111111] || [itag isEqual:@137])
+			{
+				//This will ensure the we do not query keys that are not in the `video` object's `streamURLs` and will ensure that the URL is the same was the value of specified key
+				specifiedStreamURLs[itag] = [NSURL URLWithString:@"https://www.youtube.com"];
+				continue;
+			}
+
+			specifiedStreamURLs[itag] = video.streamURLs[itag];
+		}
+
+		[[XCDYouTubeClient defaultClient] queryVideo:video streamURLsToQuery:specifiedStreamURLs options:nil cookies:nil completionHandler:^(NSDictionary * _Nullable streamURLs, NSError * _Nullable queryError, NSDictionary<id,NSError *> * _Nullable streamErrors)
+		{
+			XCTAssertNil(queryError);
+			XCTAssertNil(streamErrors);
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+
+			XCTAssertEqual(1, streamURLs.count, @"`streamURLs` should be equal to 1 since we know only 1 of the specified streams would be queried.");
+			XCTAssertNotNil(streamURLs[@22]);
+
+			[expectation fulfill];
+		}];
+	}];
+
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testQueryingWhenNoSpecifiedURLsAreInVideoObject
+{
+	/**
+	 * This video `NZzQQ1090wc` contains 24 streamURLs that are reachable ( all the URLs returned in `video.streamURLs` that are reachable)
+	 * This test ensure that when none of the  specified URLs are contained in `video.streamURLs` that we fallback to using `video.streamURLs` for querying.
+	 */
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"NZzQQ1090wc" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+
+		NSMutableDictionary *specifiedStreamURLs = [NSMutableDictionary new];
+		specifiedStreamURLs[@1111111] = [NSURL URLWithString:@"https://www.youtube.com"];
+
+		[[XCDYouTubeClient defaultClient] queryVideo:video streamURLsToQuery:specifiedStreamURLs options:nil cookies:nil completionHandler:^(NSDictionary * _Nullable streamURLs, NSError * _Nullable queryError, NSDictionary<id,NSError *> * _Nullable streamErrors)
+		{
+			XCTAssertNil(queryError);
+			XCTAssertNil(streamErrors);
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+
+			XCTAssertNotEqual(specifiedStreamURLs.count, streamURLs.count, @"`specifiedStreamURLs` should not be equal to `streamURLs` since when no streamURL is contained in `video.streamURLs` we use the `video.streamURLs` for querying. In this test we know `video.streamURLs` contains 24 objects.");
+
+			[expectation fulfill];
+		}];
+	}];
+
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+-(void) testQueryingWhenSpecifiedURLsAreEmpty
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"NZzQQ1090wc" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNotNil(video);
+		XCTAssertNil(error);
+
+		[[XCDYouTubeClient defaultClient] queryVideo:video streamURLsToQuery:@{} options:nil cookies:nil completionHandler:^(NSDictionary * _Nullable streamURLs, NSError * _Nullable queryError, NSDictionary<id,NSError *> * _Nullable streamErrors)
+		{
+			XCTAssertNil(queryError);
+			XCTAssertNil(streamErrors);
+			XCTAssertNotNil(streamURLs);
+			XCTAssertTrue([NSThread isMainThread]);
+
+			for (id key in streamURLs.allKeys)
+			{
+				XCTAssertNotNil(streamURLs[key]);
+			}
+
+			XCTAssertEqual(video.streamURLs.count, streamURLs.count, @"`streamURLs` count should be equal to `video.streamURLs` since we specified an empty array and should fallback to the `video` object's `streamURLs`. We also, know all the streamsURLs are reachable.");
+
+			[expectation fulfill];
+		}];
+	}];
+
 	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
@@ -214,7 +652,7 @@
 		XCTAssertNil(video);
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorNoStreamAvailable);
-		XCTAssertEqualObjects(error.localizedDescription, @"This video has been removed for violating YouTube's Terms of Service.");
+		XCTAssertEqualObjects(error.localizedDescription, @"This video is no longer available due to a copyright claim by Digital Rights Group Ltd.");
 		[expectation fulfill];
 	}];
 	[self waitForExpectationsWithTimeout:5 handler:nil];
@@ -243,6 +681,19 @@
 		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
 		XCTAssertEqual(error.code, XCDYouTubeErrorNoStreamAvailable);
 		XCTAssertEqualObjects(error.localizedDescription, @"Invalid parameters.");
+		[expectation fulfill];
+	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
+- (void) testEmptyResponse_offline
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"HxaM6UJpAyg" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+ 		XCTAssertNil(video);
+		XCTAssertEqualObjects(error.domain, XCDYouTubeVideoErrorDomain);
+		XCTAssertEqual(error.code, XCDYouTubeErrorEmptyResponse);
 		[expectation fulfill];
 	}];
 	[self waitForExpectationsWithTimeout:5 handler:nil];
@@ -346,6 +797,26 @@
 	[expectation performSelector:@selector(fulfill) withObject:nil afterDelay:0.2];
 	[operation cancel];
 	[self waitForExpectationsWithTimeout:1 handler:nil];
+}
+
+- (void) testCancelingOperationQueryOperation
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	__block XCDYouTubeVideoQueryOperation *operation = nil;
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"6kLq3WMV1nU" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	 {
+		XCTAssertNotNil(video);
+		
+		operation = [[XCDYouTubeClient defaultClient] queryVideo:video cookies:nil completionHandler:^(NSDictionary * _Nonnull streamURLs, NSError * _Nullable queryError, NSDictionary<id,NSError *> * _Nonnull streamErrors)
+		{
+			XCTFail();
+		}];
+		
+		[operation cancel];
+		[expectation fulfill];
+	}];
+	
+	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void) testNilCompletionHandler