+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+#### Version 2.14.0
+
+* Add new `XCDYouTubeErrorEmptyResponse` error (#479)
+* Adaptation to YouTube API change. (a048776eb0d87de869c1636e244d78ffb311adc6)
+
+#### Version 2.13.0
+
+* Deprecated `-[XCDYouTubeVideo thumbnailURL]` use `-[XCDYouTubeVideo thumbnailURLs]` (#477)
+* Exclude plist files from Swift Package Manager (#478)
+* Annotate `languageIdentifier` argument as `nullable` in `XCDYouTubeVideoOperation`
+* Add new `author`, `videoDescription` and `channelIdentifier` properties to `XCDYouTubeVideo` (#471)
+
+#### Version 2.12.0
+
+* Add the ability to specify which `streamURLs` to query via `-[XCDYouTubeClient queryVideo:video:streamURLsToQuery:options:cookies:completionHandler:]` & `XCDYouTubeVideoQueryOperation` class.
+	* `initWithVideo:streamURLsToQuery:options:cookies:` is now the designated initializer for  `XCDYouTubeVideoQueryOperation`.
+* Correctly annotate the `streamURLs` & `streamErrors` properties as `nullable`.
+
+#### Version 2.11.0
+
+* `-[XCDYouTubeClient queryVideo:video:cookies:completionHandler:]` completion handler is now correctly called on the main thread.
+* Improved video querying to handle videos that have incomplete streams. (#456)
+* Errors returned in `streamErrors` may contain the `NSLocalizedRecoverySuggestionErrorKey` key in `-[NSError userInfo]` when `NSURLErrorNetworkConnectionLost` is reported, this may indicate the file is incomplete on YouTube's server. 
+* Add logging to `XCDYouTubeVideoQueryOperation` class.
+* Improved `-[XCDYouTubeVideoQueryOperation description]`.
+
+#### Version 2.10.0
+
+* Fixed issue that caused certain videos to return error (#468)
+* Add the ability to query `XCDYouTubeVideo` for reachable streams via `-[XCDYouTubeClient queryVideo:video:cookies:completionHandler:]` & `XCDYouTubeVideoQueryOperation` class
+* Add new `videoIdentifiers` property in `XCDYouTubeVideo` (#290)
+
+#### Version 2.9.0
+
+* Add the ability to use custom regular expression patterns via `-[XCDYouTubeClient getVideoWithIdentifier:cookies:customPatterns:completionHandler:]` & `initWithVideoIdentifier:cookies:languageIdentifier:customPatterns` (#463, #199)
+* Add new `viewCount` property in `XCDYouTubeVideo` (#460)
+* Silence warning about deprecated implementation (#450)
+
 #### Version 2.8.3
 
 * Adaptation to YouTube API change. (#458)